docs: update README with detailed sections on foundation model selection and fine-tuning memory optimization
- Expanded 3.1 with guidance on choosing foundation models (proprietary vs. open-source, evaluation criteria, cost, and licensing)
- Added detailed memory lifecycle breakdown during fine-tuning
- Documented memory optimization techniques including:
  - Gradient checkpointing
  - Gradient accumulation
  - Mixed precision training
  - Efficient optimizers (Adafactor, 8-bit, paged)
  - Offloading strategies (ZeRO, FSDP)
  - Zeroth-order optimization (MeZO, DiZO, SubZero)
  - FlashAttention-2 integration example
- Included benchmarking examples and code snippets for reproducibility

+++ finetune/experiments/optimization/benchmark_v1.0.py
@@ -33,7 +33,7 @@ tokenized_eval.set_format(type="torch", columns=["input_ids", "attention_mask", 
 
 # Define models to benchmark
 models_to_benchmark = [
-    (model_name, AutoModelForCausalLM.from_pretrained(model_name)),
+    (model_name, AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=torch.bfloat16)),
 ]
 
 # Set up training arguments
